Add files via upload
What Satoshi originally dreamed of before government interference and manipulation, before the super rich stockpiled to create it unobtainable for the regular man. The original Bitcoin now on the ERC network we bring Bitcoin ($BC), with 21 million circulating supply, a fair launched ERC token to commemorate the original Bitcoin how Satoshi envisioned it.


We want to use Trustwallet effectively. Please allow our logo to be uploaded.