New PR needed for Migration of PR (#12009) (#24452)
* Adding PEEPS to TrustWallet

PEEPS is a community-driven coin built by people within that very same community. It will host an ever-adapting ecosystem built with the visibility and security of the ETH network. The transparency and welcoming community will both allow organic growth and the promotion of ideas to improve the existing crypto landscape.

* Adding PEEPS to TrustWallet

PEEPS is a community-driven coin built by people within that very same community. It will host an ever-adapting ecosystem built with the visibility and security of the ETH network. The transparency and welcoming community will both allow organic growth and the promotion of ideas to improve the existing crypto landscape.
Commit directly to the master branch.

* Update info.json

* Add files via upload

* Delete logo.png

* Update info.json

* Delete logo.png

* Create info.json

* Create logo.png

* Update info.json

* Delete logo.png

---------

Co-authored-by: Nazar <85495047+cryptomanz@users.noreply.github.com>
Co-authored-by: cryptomanz1 <85495047+cryptomanz1@users.noreply.github.com>