ADD TPT logo and info
https://bscscan.com/tx/0x4bf3b74738e0cb7e7cffdeb3ba21a45e2728e287e547c43ec44301bcdaacaaf3

This token contract originally had an agent upgrade function, but has given up the contract ownership and has passed a security audit https://github.com/cyberscope-io/audits/blob/main/dis/audit.pdf https://www.coinscope.co/coin/dis/audit https://twitter.com/cyberscope_io/status/1668656345178861568 Another important point is that we use the real-name company to issue the token Miningtw Technology Co., Ltd. https://www.twincn.com/item.aspx?no=90324590 And it is stated in the white paper that this DIS has nothing to do with Walt Disney. It is not issued or related to them. It is a meme coin with similar names in different fields.

//re:
Token is detected to have a proxy contract, which is a potentially dangerous configuration. Therefore, unless there is a formal security audit that proves its safety, it cannot be listed. However, the TPT hash value you submitted can be used in the next logo submission.//