Add Blackjack.fun coin (JACK) logo and info
JACK is a token made by the creators of blackjack.fun crypto casino. JACK was born to enforce a bridge between the platform, the community and the blackjack tournaments. This concept became a reality to further develop an opportunity for the community and to literally have a part of the project.

Please update the logo and JACK token information. Thank you.