Add ORDS token info
Our token is already listed on several CEXs.
- https://www.gate.io/trade/ORDS_USDT
- https://www.htx.com/en-us/trade/ords_usdt
- https://www.mexc.com/exchange/ORDS_USDT
- https://poloniex.com/trade/ORDS_USDT
- https://www.bitget.com/spot/ORDSUSDT
- https://www.bitmart.com/trade/en-US?symbol=ORDS_USDT

ORDS Token is an interoperable utility token meticulously crafted for seamless operation across Bitcoin's native layer and the Ethereum network. With a finite supply, $ORDS is the linchpin of Ordiswap's developmental prowess, facilitating community-driven decisions through voting and governance.

Please let we know if we pay the FEE(5BNB) and our token can be listed on trustwallet.
Thanks.