AlienForm (A4M) Listing Request
Requesting to have our token's information listed on Trust Wallet. We are a new project founded by the community of the Shibnobi (Shinja) project. The developer of Shibnobi scammed and deserted the project, and we have taken over by launching AlienForm and airdropping every holder of Shinja v2 our A4M token.