We add Syndicatus Logo and json in 256x256 png 2
We are adding Syndicatus Token (SYND) logo and the info.json folder which contains all the details about our token project. This is our first time doing this, and we're only hoping that we did it right. Please do guide us on to correcting any wrong doing, which we don't think so, so far.