FQS
Greetings and Regards
Our cryptocurrency project is listed in the coin market with the abbreviation FQS after migrating from version 1 to 2. Currently, it is listed and priced in Trust Wallet with the abbreviation FQS, but it does not have a logo.
The logo of version 2 has been changed and is also visible in Coin Market. Please check our information and display our logo along with the price in Trust Wallet.
Thanks
FQS team