Create EYE NETWORK
Hi @Cryptocool1 @cryptomanz1 , [ EYE Network (BEP20) #27846 ]

I'm so sorry again to bother you guys, but it's been more than 24 hours. Please advise me if I need to create a new pull request or if you guys will check it here. I'm mentioning again the old PR that I paid for, and now it's listed on CMC already.
Please try to update the logo for the EYE NETWORK token. It's been a few days since Eye was listed over CMC, and we are waiting for a response from you.
https://coinmarketcap.com/currencies/eye-network/

EYE Network (BEP20) #27846

Thank you