Create 0xC016ADf23B86503929738E04E6fBF0d673649784
"description": "Dogestine is a unique meme coin that combines the fun of Dogecoin with a deeper message of hope and peace. The Doge dog proudly wears a Palestinian scarf, symbolized by the peace dove, representing unity and reconciliation. Dogestine is more than just a token – it's a statement of solidarity and global connectedness. With every transaction, 1% of the volume is donated to charitable causes, helping those in need and promoting positive change. Support peace and be part of a movement for good."