Update info.json
This commit updates the token information and logo for MeshchainCoin (MCC) in the Trust Wallet repository. 

We are submitting this request because the MCC token price and token logo are currently not displaying correctly on Trust Wallet. To resolve this, we are making the following updates:
- Added the MeshchainCoin (MCC) logo as Logo.png in the assets folder to ensure it displays properly in the wallet.
- Updated the token's Social Media & Links in info.json, along with other relevant details.
- Ensured that the logo meets Trust Wallet's size and resolution guidelines for proper display.

This update will allow Trust Wallet users to view and manage MeshchainCoin (MCC) with the updated logo and accurate information, including the token's price and relevant links.