logo.png
Adding ZoZo Token to Binance Smart Chain
This is a meme token created to support the community and fight fake tokens. 
ZoZo is designed to bring hope to meme coin enthusiasts with its strong community-driven focus and rewards. 

The logo and information for ZoZo token have been added to the respective directories for proper display and usage on Trust Wallet.

- **Token Address:** 0x022f34755C44c88a522f1cFD60E0EfD831F958a9
- **Logo Path:** /blockchains/smartchain/assets/0x022f34755C44c88a522f1cFD60E0EfD831F958a9/logo.png

This PR includes all the necessary details for the ZoZo token, including the correct logo image and token information.