Create 0x296f66f16fc54a04d748821d87b440ea65eeaf02
This token is developed as a game development token under the name 'Tether USDT'. It is designed for in-game purchases and to be used at the same value and liquidity as the original Tether (USDT). This helps prevent exchange rate losses and ensures stable USD-based investment and development tracking. The token supports both digital asset protection and internal game economy.