Update info.json
I have manually checked the number of holding addresses of the token contract on BscScan, and there are currently 14 holding addresses. The holding distribution is relatively reasonable, without extreme concentration, and meets TrustWallet's basic requirements for community holding distribution. The holding addresses are all active users, and the tokens are not all concentrated in a single address. Please refer to the audit team, thank you!