Add CYBRO dapp logo
Dapp name for the listing
CYBRO

Dapp URL
https://cybro.io/

Short description

CYBRO is a multichain AI-powered earn marketplace

Long Description

CYBRO is an earn marketplace where users can choose investment tools based on desired returns and risk levels. CYBRO's main feature is a set of investment Vaults (Strategies) for yield and/or point farming.

CYBRO acts as a multichain yield aggregator on Blast, Arbitrum, and Base. It enhances this offering with its own combined with its One-Click strategies, providing users with a streamlined and efficient way to optimize their crypto investments.

The killer features of CYBRO are: the AI Broker, smart notifications, and a unique set of DeFi investment vehicles.