Update info.json
Hello Trust Wallet Team,

I am submitting the token MomDadX (MMDX) for listing. Currently, the token has 2 holders as it is in the early stages of development. We are actively working on expanding our community and increasing the number of holders.

We understand that Trust Wallet requires a minimum of 10,000 holders and 15,000 transactions for listing approval. While we have not yet met these criteria, we hope you will consider our submission and provide any guidance or feedback to help us move forward.

Thank you for your time and consideration.

Best regards,
Hamid