Request to list a new token
Fast, Scalable Stablecoin Transactions!  
wUSDT is a wrapped version of Tether (USDT) on the Solana blockchain, bringing the world’s most popular stablecoin to one of the fastest and most cost-efficient networks in crypto. With low fees, instant settlements, and high scalability, wUSDT on Solana is perfect for DeFi, trading, and cross-chain transactions.